use adaptive_hash table optimization only for single-user/app use cases

+++ kam1n0/kam1n0-clone/src/main/java/ca/mcgill/sis/dmas/kam1n0/problem/clone/detector/kam/index/LshAdaptiveDupIndexCasandra.java
@@ -28,13 +28,7 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
-import com.datastax.driver.core.ResultSet;
-import com.datastax.driver.core.SimpleStatement;
-import com.datastax.driver.core.Statement;
-import com.datastax.driver.core.querybuilder.BuiltStatement;
-import org.apache.hadoop.yarn.webapp.hamlet.Hamlet;
 import org.apache.spark.api.java.JavaRDD;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -56,22 +50,35 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 
 	private String databaseName;
 	private CassandraInstance cassandraInstance;
+	private boolean singleUserApplication;
 
 	private static Logger logger = LoggerFactory.getLogger(LshAdaptiveDupIndexCasandra.class);
 
+	/**
+	 *
+	 * @param sparkInstance
+	 * @param cassandraInstance
+	 * @param databaseName
+	 * @param singleUserApplication   When set, adaptive_hash table is partitioned by hash ID instead of applicationID,
+	 * 	  							  which is more efficient and does not require adaptive_hash_view anymore.
+	 * 	  							  This may only be set for single-user/app cases, and must be the same value when
+	 * 	  							  creating and re-using a database. Note: 'appliation ID' ('rid0' column) is still
+	 * 	  							  stored in the DB (as a clustering key) for code compatibility.
+	 */
 	public LshAdaptiveDupIndexCasandra(SparkInstance sparkInstance, CassandraInstance cassandraInstance,
-			String databaseName) {
+			String databaseName, boolean singleUserApplication) {
 		super(sparkInstance);
 		this.databaseName = databaseName;
 		this.cassandraInstance = cassandraInstance;
+		this.singleUserApplication = singleUserApplication;
 	}
 
 	// classes:
 	public static final String _ADAPTIVE_HASH = "ADAPTIVE_HASH".toLowerCase();
-	//public static final String _ADAPTIVE_HASH_VIEW = _ADAPTIVE_HASH + "_view";
+	public static final String _ADAPTIVE_HASH_VIEW = _ADAPTIVE_HASH + "_view";
 
 	// properties:
-	public static final String _APP_ID = "rid0";
+	public static final String _APP_ID = "rid0";	// always the same and irrelevant when if 'singleUserApplication'
 	public static final String _ADAPTIVE_HASH_HASHID = "hashid";
 	public static final String _ADAPTIVE_HASH_IND = "ind";
 	public static final String _ADAPTIVE_HASH_FULLKEY = "full_key";
@@ -85,6 +92,9 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 						+ "replication = {'class':'SimpleStrategy', 'replication_factor':1} "
 						+ " AND durable_writes = true;");
 
+				String primaryKey = singleUserApplication ? "(" + _ADAPTIVE_HASH_HASHID + ")," + _APP_ID
+						: "(" + _APP_ID + ")," + _ADAPTIVE_HASH_HASHID;
+
 				String createTableStatement = "create table if not exists " + databaseName + "." + _ADAPTIVE_HASH + " (" //
 						+ _APP_ID + " bigint," //
 						+ _ADAPTIVE_HASH_HASHID + " bigint, " //
@@ -92,23 +102,34 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 						+ _ADAPTIVE_HASH_FULLKEY + " blob, " //
 						+ _ADAPTIVE_HASH_SHARED_INFO + " text, " //
 						+ _ADAPTIVE_HASH_VIDS + " set<text>, " //
-//						+ " PRIMARY KEY ((" + _APP_ID + ")," + _ADAPTIVE_HASH_HASHID + ")" //
-						+ " PRIMARY KEY ((" + _APP_ID + "," + _ADAPTIVE_HASH_HASHID + ")," + _ADAPTIVE_HASH_IND + ")" //
+						+ " PRIMARY KEY (" + primaryKey + ")"
 						+ ");";
 				logger.info("Creating table {}.{} with {}", this.databaseName, _ADAPTIVE_HASH, createTableStatement);
 				session.execute(createTableStatement);
 
-//				String createViewStatement = "CREATE MATERIALIZED VIEW " + databaseName + "." + _ADAPTIVE_HASH_VIEW + " AS "//
-//						+ "SELECT * FROM " + databaseName + "." + _ADAPTIVE_HASH + " "//
-//						+ "WHERE " + _APP_ID + " IS NOT NULL AND " + _ADAPTIVE_HASH_HASHID + " IS NOT NULL AND "
-//						+ _ADAPTIVE_HASH_IND + " IS NOT NULL " + "PRIMARY KEY((" + _APP_ID + ", "
-//						+ _ADAPTIVE_HASH_HASHID + "), " + _ADAPTIVE_HASH_IND
-//						+ ")" + tombstoneManagementView + ";";
-//				logger.info("Creating view {}.{} with {}", this.databaseName, _ADAPTIVE_HASH_VIEW, createViewStatement);
-//				session.execute(createViewStatement);
+				if (!singleUserApplication) {
+					String createViewStatement = "CREATE MATERIALIZED VIEW " + databaseName + "." + _ADAPTIVE_HASH_VIEW + " AS "//
+							+ "SELECT * FROM " + databaseName + "." + _ADAPTIVE_HASH + " "//
+							+ "WHERE " + _APP_ID + " IS NOT NULL AND " + _ADAPTIVE_HASH_HASHID + " IS NOT NULL AND "
+							+ _ADAPTIVE_HASH_IND + " IS NOT NULL " + "PRIMARY KEY((" + _APP_ID + ", "
+							+ _ADAPTIVE_HASH_HASHID + "), " + _ADAPTIVE_HASH_IND
+							+ ");";
+					logger.info("Creating view {}.{} with {}", this.databaseName, _ADAPTIVE_HASH_VIEW, createViewStatement);
+					session.execute(createViewStatement);
+				}
+
 			});
 		} else {
 			logger.info("Found table {}.{}", this.databaseName, _ADAPTIVE_HASH);
+
+			boolean hasView = cassandraInstance.checkColumnFamilies(
+					this.sparkInstance.getConf(), this.databaseName, _ADAPTIVE_HASH_VIEW);
+
+			if (hasView == singleUserApplication) {
+				throw new RuntimeException(databaseName + "." + _ADAPTIVE_HASH + " was created as a "
+						+ (hasView ? "multi" : "single") + "-user application DB, but tried to be opened as a "
+						+ (singleUserApplication ? "single" : "multi") + "-user application DB");
+			}
 		}
 	}
 
@@ -162,8 +183,7 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 							session.executeAsync(QueryBuilder.update(databaseName, _ADAPTIVE_HASH)
 									.with(addAll(_ADAPTIVE_HASH_VIDS, vals)) //
 									.where(eq(_APP_ID, rid)) //
-									.and(eq(_ADAPTIVE_HASH_HASHID, vec.hashId))
-									.and(eq(_ADAPTIVE_HASH_IND, vec.ind)));
+									.and(eq(_ADAPTIVE_HASH_HASHID, vec.hashId)));
 							return null;
 						}
 					}).filter(vec -> vec != null).collect(Collectors.toList());
@@ -270,13 +290,14 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 		else
 			selection = selectAllInfo;
 
+		String dbToQuery = singleUserApplication ? _ADAPTIVE_HASH : _ADAPTIVE_HASH_VIEW;
+
 		if (sparkInstance.localMode) {
 			List<VecEntry<T, K>> collected = this.cassandraInstance.doWithSessionWithReturn(sparkInstance.getConf(),
 					session -> {
 						return hashIds.parallelStream().map(hid -> {
 							return session.execute(QueryBuilder.select(selection)
-	//								.from(databaseName, _ADAPTIVE_HASH_VIEW).where(eq(_APP_ID, rid)) //
-									.from(databaseName, _ADAPTIVE_HASH).where(eq(_APP_ID, rid))
+									.from(databaseName, dbToQuery).where(eq(_APP_ID, rid))
 									.and(eq(_ADAPTIVE_HASH_HASHID, hid))).one();
 						}).filter(row -> row != null).map(row -> {
 							VecEntry<T, K> vec = new VecEntry<T, K>();
@@ -287,10 +308,10 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 							bb.get(vec.fullKey);
 							if (!excludeBlockIds) {
 								Set<String> vids = row.getSet(3, String.class);
-//								int approxSize = vids.stream().mapToInt(s -> s.length()).sum();
-//								if ( approxSize > 100 * 1048576 ) {
-//									logger.warn("kam182 vids {} {}", vec.hashId, approxSize);
-//								}
+								int approxSize = vids.stream().mapToInt(s -> s.length()).sum();
+								if ( approxSize > 10 * 1048576 ) {
+									logger.warn("kam182 vids {} {}", vec.hashId, approxSize);
+								}
 								List<T> ls = vids.stream().map(VecInfo::<T>deSerialize).collect(Collectors.toList());
 								if (filter != null) {
 									ls = filter.apply(ls);
@@ -309,7 +330,7 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 		} else {
 			return javaFunctions(this.sparkInstance.getContext())
 					//
-					.cassandraTable(databaseName, _ADAPTIVE_HASH)
+					.cassandraTable(databaseName, dbToQuery)
 					//
 					.select(selection)//
 					.where(_APP_ID + " = ? AND " + _ADAPTIVE_HASH_HASHID + " in ?", rid, hashIds)//
@@ -357,8 +378,12 @@ public class LshAdaptiveDupIndexCasandra<T extends VecInfo, K extends VecInfoSha
 	public void clear(long rid) {
 		try {
 			this.cassandraInstance.doWithSession(sess -> {
-				sess.executeAsync(QueryBuilder.delete().from(databaseName, _ADAPTIVE_HASH)//
-						.where(eq(_APP_ID, rid)));
+				if (singleUserApplication) {
+					sess.executeAsync(QueryBuilder.truncate(databaseName, _ADAPTIVE_HASH));
+				} else {
+					sess.executeAsync(QueryBuilder.delete().from(databaseName, _ADAPTIVE_HASH)//
+							.where(eq(_APP_ID, rid)));
+				}
 			});
 		} catch (Exception e) {
 			logger.error("Failed to delete the index.", e);
